Resolve "WASP Utility Deployment"

Closes #39

See merge request !45

+++ wasputils/ddiselect.cpp
@@ -74,7 +74,7 @@ int main(int argc, char** argv) {
 
     std::vector<decltype(parser.root())> select_from_node;
     select_from_node.push_back(parser.root());
-    for( size_t j = 3; j < argc; ++j)
+    for( int j = 3; j < argc; ++j)
     {
         std::stringstream select_statement_errors;
         std::string select_statement = argv[j];
